requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.10
Django==4.0.1
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
fire==0.4.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-dateutil==1.5
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
termcolor==1.1.0
urllib3==1.26.8
whitenoise==5.3.0
